wcf made switchable
