+++ CefSharp.BrowserSubprocess/CefSharp.BrowserSubprocess.csproj
@@ -91,6 +91,7 @@
     <Compile Include="CefGpuProcess.cs" />
     <Compile Include="CefRenderProcess.cs" />
     <Compile Include="CefSubProcess.cs" />
+    <Compile Include="CefWcfRenderProcess.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
