+++ CefSharp.Core/Cef.h
@@ -33,6 +33,7 @@ namespace CefSharp
     private:
         static Object^ _sync;
 
+        static bool _wcfEnabled = false;
         static bool _initialized = false;
         static HashSet<IDisposable^>^ _disposables;
 
@@ -72,6 +73,14 @@ namespace CefSharp
             msclr::lock l(_sync);
 
             _disposables->Remove(item);
+        }
+
+        static property bool WcfEnabled
+        {
+            bool get()
+            {
+                return _wcfEnabled;
+            }
         }
 
         /// <summary>Gets a value that indicates whether CefSharp is initialized.</summary>
@@ -182,6 +191,10 @@ namespace CefSharp
                 }
 
                 _initialized = success;
+                if (success)
+                {
+                    _wcfEnabled = cefSettings->WcfEnabled;
+                }
 
                 if (_initialized && shutdownOnProcessExit)
                 {
