+++ CefSharp/Internals/CefSharpArguments.cs
@@ -1,0 +1,7 @@
+﻿namespace CefSharp.Internals
+{
+    public static class CefSharpArguments
+    {
+        public const string WcfEnabledArgument = "wcf-enabled";
+    }
+}
